LayoutTests/crypto/subtle/aes-kw-generate-key.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test generating an AES key with length 128 using AES-KW algorithm.");

jsTestIsAsync = true;

debug("Generating a key...");
crypto.subtle.generateKey({name: "aes-kw", length: 128}, true, ["unwrapKey", "wrapKey"]).then(function(result) {
    key = result;

    shouldBe("key.type", "'secret'");
    shouldBe("key.extractable", "true");
    shouldBe("key.algorithm.name", "'AES-KW'");
    shouldBe("key.algorithm.length", "128");
    shouldBe("key.usages", "['unwrapKey', 'wrapKey']");

    finishJSTest();
});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
