Safari-609.1.11.1/LayoutTests/crypto/subtle/aes-kw-generate-key-expected.txt

Test generating an AES key with length 128 using AES-KW algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating a key...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'AES-KW'
PASS key.algorithm.length is 128
PASS key.usages is ['unwrapKey', 'wrapKey']
PASS successfullyParsed is true

TEST COMPLETE
